[FLINK-28256] Move the write and prepareCommit logic of AbstractTableWrite to FileStoreWrite
