+++ flink-table-store-core/src/main/java/org/apache/flink/table/store/file/operation/AppendOnlyFileStoreWrite.java
@@ -35,6 +35,7 @@ import org.apache.flink.table.store.file.utils.RecordReaderIterator;
 import org.apache.flink.table.store.file.utils.RecordWriter;
 import org.apache.flink.table.store.file.utils.SnapshotManager;
 import org.apache.flink.table.store.format.FileFormat;
+import org.apache.flink.table.store.table.sink.WriteFunction;
 import org.apache.flink.table.store.table.source.Split;
 import org.apache.flink.table.types.logical.RowType;
 
@@ -69,8 +70,9 @@ public class AppendOnlyFileStoreWrite extends AbstractFileStoreWrite<RowData> {
             FileStorePathFactory pathFactory,
             SnapshotManager snapshotManager,
             FileStoreScan scan,
-            CoreOptions options) {
-        super(snapshotManager, scan);
+            CoreOptions options,
+            WriteFunction<RowData> writeFunction) {
+        super(snapshotManager, scan, writeFunction);
         this.read = read;
         this.schemaId = schemaId;
         this.rowType = rowType;
