+++ flink-table-store-core/src/main/java/org/apache/flink/table/store/table/AppendOnlyFileStoreTable.java
@@ -30,18 +30,14 @@ import org.apache.flink.table.store.file.schema.SchemaManager;
 import org.apache.flink.table.store.file.schema.TableSchema;
 import org.apache.flink.table.store.file.utils.FileStorePathFactory;
 import org.apache.flink.table.store.file.utils.RecordReader;
-import org.apache.flink.table.store.file.utils.RecordWriter;
-import org.apache.flink.table.store.table.sink.AbstractTableWrite;
-import org.apache.flink.table.store.table.sink.SinkRecord;
 import org.apache.flink.table.store.table.sink.SinkRecordConverter;
 import org.apache.flink.table.store.table.sink.TableWrite;
+import org.apache.flink.table.store.table.sink.TableWriteImpl;
 import org.apache.flink.table.store.table.source.AppendOnlySplitGenerator;
 import org.apache.flink.table.store.table.source.Split;
 import org.apache.flink.table.store.table.source.SplitGenerator;
 import org.apache.flink.table.store.table.source.TableRead;
 import org.apache.flink.table.store.table.source.TableScan;
-import org.apache.flink.types.RowKind;
-import org.apache.flink.util.Preconditions;
 
 import java.io.IOException;
 
@@ -108,17 +104,7 @@ public class AppendOnlyFileStoreTable extends AbstractFileStoreTable {
     public TableWrite newWrite() {
         SinkRecordConverter recordConverter =
                 new SinkRecordConverter(store.options().bucket(), tableSchema);
-        return new AbstractTableWrite<RowData>(store.newWrite(), recordConverter) {
-            @Override
-            protected void writeSinkRecord(SinkRecord record, RecordWriter<RowData> writer)
-                    throws Exception {
-                Preconditions.checkState(
-                        record.row().getRowKind() == RowKind.INSERT,
-                        "Append only writer can not accept row with RowKind %s",
-                        record.row().getRowKind());
-                writer.write(record.row());
-            }
-        };
+        return new TableWriteImpl<>(store.newWrite(), recordConverter);
     }
 
     @Override
