+++ flink-table-store-core/src/main/java/org/apache/flink/table/store/table/ChangelogValueCountFileStoreTable.java
@@ -34,10 +34,11 @@ import org.apache.flink.table.store.file.schema.TableSchema;
 import org.apache.flink.table.store.file.utils.FileStorePathFactory;
 import org.apache.flink.table.store.file.utils.RecordReader;
 import org.apache.flink.table.store.file.utils.RecordWriter;
-import org.apache.flink.table.store.table.sink.MemoryTableWrite;
 import org.apache.flink.table.store.table.sink.SinkRecord;
 import org.apache.flink.table.store.table.sink.SinkRecordConverter;
 import org.apache.flink.table.store.table.sink.TableWrite;
+import org.apache.flink.table.store.table.sink.TableWriteImpl;
+import org.apache.flink.table.store.table.sink.WriteFunction;
 import org.apache.flink.table.store.table.source.KeyValueTableRead;
 import org.apache.flink.table.store.table.source.MergeTreeSplitGenerator;
 import org.apache.flink.table.store.table.source.SplitGenerator;
@@ -72,7 +73,35 @@ public class ChangelogValueCountFileStoreTable extends AbstractFileStoreTable {
                         tableSchema.logicalBucketKeyType(),
                         tableSchema.logicalRowType(),
                         countType,
-                        mergeFunction);
+                        mergeFunction,
+                        new WriteFunction<KeyValue>() {
+                            private final KeyValue kv = new KeyValue();
+
+                            @Override
+                            public void write(SinkRecord record, RecordWriter<KeyValue> writer)
+                                    throws Exception {
+                                switch (record.row().getRowKind()) {
+                                    case INSERT:
+                                    case UPDATE_AFTER:
+                                        kv.replace(
+                                                record.row(),
+                                                RowKind.INSERT,
+                                                GenericRowData.of(1L));
+                                        break;
+                                    case UPDATE_BEFORE:
+                                    case DELETE:
+                                        kv.replace(
+                                                record.row(),
+                                                RowKind.INSERT,
+                                                GenericRowData.of(-1L));
+                                        break;
+                                    default:
+                                        throw new UnsupportedOperationException(
+                                                "Unknown row kind " + record.row().getRowKind());
+                                }
+                                writer.write(kv);
+                            }
+                        });
     }
 
     @Override
@@ -122,29 +151,7 @@ public class ChangelogValueCountFileStoreTable extends AbstractFileStoreTable {
     public TableWrite newWrite() {
         SinkRecordConverter recordConverter =
                 new SinkRecordConverter(store.options().bucket(), tableSchema);
-        return new MemoryTableWrite<KeyValue>(store.newWrite(), recordConverter, store.options()) {
-
-            private final KeyValue kv = new KeyValue();
-
-            @Override
-            protected void writeSinkRecord(SinkRecord record, RecordWriter<KeyValue> writer)
-                    throws Exception {
-                switch (record.row().getRowKind()) {
-                    case INSERT:
-                    case UPDATE_AFTER:
-                        kv.replace(record.row(), RowKind.INSERT, GenericRowData.of(1L));
-                        break;
-                    case UPDATE_BEFORE:
-                    case DELETE:
-                        kv.replace(record.row(), RowKind.INSERT, GenericRowData.of(-1L));
-                        break;
-                    default:
-                        throw new UnsupportedOperationException(
-                                "Unknown row kind " + record.row().getRowKind());
-                }
-                writer.write(kv);
-            }
-        };
+        return new TableWriteImpl<>(store.newWrite(), recordConverter);
     }
 
     @Override
