+++ flink-table-store-core/src/main/java/org/apache/flink/table/store/table/ChangelogWithKeyFileStoreTable.java
@@ -36,11 +36,12 @@ import org.apache.flink.table.store.file.schema.TableSchema;
 import org.apache.flink.table.store.file.utils.FileStorePathFactory;
 import org.apache.flink.table.store.file.utils.RecordReader;
 import org.apache.flink.table.store.file.utils.RecordWriter;
-import org.apache.flink.table.store.table.sink.MemoryTableWrite;
 import org.apache.flink.table.store.table.sink.SequenceGenerator;
 import org.apache.flink.table.store.table.sink.SinkRecord;
 import org.apache.flink.table.store.table.sink.SinkRecordConverter;
 import org.apache.flink.table.store.table.sink.TableWrite;
+import org.apache.flink.table.store.table.sink.TableWriteImpl;
+import org.apache.flink.table.store.table.sink.WriteFunction;
 import org.apache.flink.table.store.table.source.KeyValueTableRead;
 import org.apache.flink.table.store.table.source.MergeTreeSplitGenerator;
 import org.apache.flink.table.store.table.source.SplitGenerator;
@@ -102,16 +103,39 @@ public class ChangelogWithKeyFileStoreTable extends AbstractFileStoreTable {
                 throw new UnsupportedOperationException("Unsupported merge engine: " + mergeEngine);
         }
 
+        CoreOptions options = new CoreOptions(conf);
+        SequenceGenerator sequenceGenerator =
+                options.sequenceField()
+                        .map(field -> new SequenceGenerator(field, schema().logicalRowType()))
+                        .orElse(null);
         this.store =
                 new KeyValueFileStore(
                         schemaManager,
                         tableSchema.id(),
-                        new CoreOptions(conf),
+                        options,
                         tableSchema.logicalPartitionType(),
                         addKeyNamePrefix(tableSchema.logicalBucketKeyType()),
                         addKeyNamePrefix(tableSchema.logicalTrimmedPrimaryKeysType()),
                         rowType,
-                        mergeFunction);
+                        mergeFunction,
+                        new WriteFunction<KeyValue>() {
+                            private final KeyValue kv = new KeyValue();
+
+                            @Override
+                            public void write(SinkRecord record, RecordWriter<KeyValue> writer)
+                                    throws Exception {
+                                long sequenceNumber =
+                                        sequenceGenerator == null
+                                                ? KeyValue.UNKNOWN_SEQUENCE
+                                                : sequenceGenerator.generate(record.row());
+                                writer.write(
+                                        kv.replace(
+                                                record.primaryKey(),
+                                                sequenceNumber,
+                                                record.row().getRowKind(),
+                                                record.row()));
+                            }
+                        });
     }
 
     private RowType addKeyNamePrefix(RowType type) {
@@ -190,30 +214,7 @@ public class ChangelogWithKeyFileStoreTable extends AbstractFileStoreTable {
     public TableWrite newWrite() {
         SinkRecordConverter recordConverter =
                 new SinkRecordConverter(store.options().bucket(), tableSchema);
-        SequenceGenerator sequenceGenerator =
-                store.options()
-                        .sequenceField()
-                        .map(field -> new SequenceGenerator(field, schema().logicalRowType()))
-                        .orElse(null);
-        return new MemoryTableWrite<KeyValue>(store.newWrite(), recordConverter, store.options()) {
-
-            private final KeyValue kv = new KeyValue();
-
-            @Override
-            protected void writeSinkRecord(SinkRecord record, RecordWriter<KeyValue> writer)
-                    throws Exception {
-                long sequenceNumber =
-                        sequenceGenerator == null
-                                ? KeyValue.UNKNOWN_SEQUENCE
-                                : sequenceGenerator.generate(record.row());
-                writer.write(
-                        kv.replace(
-                                record.primaryKey(),
-                                sequenceNumber,
-                                record.row().getRowKind(),
-                                record.row()));
-            }
-        };
+        return new TableWriteImpl<>(store.newWrite(), recordConverter);
     }
 
     @Override
